requirements.txt
watchdog
requests
python-dotenv
bencodepy
tzlocal
pyyaml